--- src/main/java/com/project/core/commands/CreateApplicationCommand.java
@@ -1,13 +1,0 @@
-package com.project.core.commands;
-
-import lombok.Builder;
-import lombok.Data;
-import org.axonframework.modelling.command.TargetAggregateIdentifier;
-@Data
-@Builder
-//TODO refactoring this class(delete unused code) and sort classes by package(application, user)
-public class CreateApplicationCommand {
-  @TargetAggregateIdentifier
-  private String clientId;
-  private String secret;
-}
