+++ src/main/java/com/project/core/commands/user/GenerateOneTimeCodeUserProfileCommand.java
@@ -1,17 +1,13 @@
-package com.project.core.commands;
+package com.project.core.commands.user;
 
 import lombok.Builder;
 import lombok.Data;
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
-import java.util.Date;
 @Data
 @Builder
-public class CreateUserProfileCommand {
+public class GenerateOneTimeCodeUserProfileCommand {
   @TargetAggregateIdentifier
   private String userId;
-  private String userName;
   private String passwordHash;
-  private String userStatus;
-  private Date createdAt;
 }
