--- src/main/java/com/project/core/events/ApplicationLoggedInEvent.java
@@ -1,11 +1,0 @@
-package com.project.core.events;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ApplicationLoggedInEvent {
-  private String clientId;
-  private String refreshToken;
-}
