--- src/main/java/com/project/core/events/ApplicationRegisteredEvent.java
@@ -1,11 +1,0 @@
-package com.project.core.events;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ApplicationRegisteredEvent {
-  private String clientId;
-  private String code;
-}
