+++ src/main/java/com/project/core/events/app/ApplicationCreatedEvent.java
@@ -1,11 +1,11 @@
-package com.project.core.queries;
+package com.project.core.events.app;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class CheckLoginDataQuery {
+public class ApplicationCreatedEvent {
   private String clientId;
   private String secret;
 }
