+++ src/main/java/com/project/core/events/user/UserProfileCreatedEvent.java
@@ -1,0 +1,15 @@
+package com.project.core.events.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.Date;
+@Data
+@Builder
+public class UserProfileCreatedEvent {
+  private String userId;
+  private String userName;
+  private String passwordHash;
+  private String userStatus;
+  private Date createdAt;
+}
