+++ src/main/java/com/project/core/queries/app/CheckLoginDataQuery.java
@@ -1,11 +1,11 @@
-package com.project.core.events;
+package com.project.core.queries.app;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class ApplicationCreatedEvent {
+public class CheckLoginDataQuery {
   private String clientId;
   private String secret;
 }
